// src/@types/components/blocks/Navigation/types.ts
export interface NavigationProperties {
	activeLink: string
}

export interface NavigationLeftMenuItem {
	id: string
	img: string
	href: string
}

export interface NavigationMenuItem {
	id: string
	title: string
	href: string
}

export interface NavigationLanguageItem {
	id: string
	language: string
}